fix: fix all mistakes
