+++ news-JS/src/components/view/appView.ts
@@ -1,23 +1,7 @@
+import { DataType } from '../types/types';
 import News from './news/news';
-// import Sources from './sources/sources';
 import { showByAlph } from './sources/sources';
-export type DataType = {
-    articles: articleType[];
-    sources: sourceType[];
-};
-export type articleType = {
-    author: string;
-    publishedAt: string;
-    source: sourceType;
-    title: string;
-    url: string;
-    description: string;
-    urlToImage: string;
-};
-export type sourceType = {
-    id: string;
-    name: string;
-};
+
 export class AppView {
     news: News;
     // sources: Sources;
