+++ news-JS/src/components/view/news/news.css
@@ -1,4 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Merriweather&display=swap');
+
 .news__item {
     display: flex;
     flex-direction: column;
@@ -9,8 +10,8 @@
     line-height: 1.4;
     font-family: Arial, sans-serif;
     border-radius: 5px;
-    overflow: hidden;
     box-shadow: rgba(0, 0, 0, 0.07) 0px 1px 2px, rgba(0, 0, 0, 0.07) 0px 2px 4px, rgba(0, 0, 0, 0.07) 0px 4px 8px, rgba(0, 0, 0, 0.07) 0px 8px 16px, rgba(0, 0, 0, 0.07) 0px 16px 32px, rgba(0, 0, 0, 0.07) 0px 32px 64px;
+    overflow: hidden;
 }
 
 .news__item:hover .news__meta-photo {
@@ -41,6 +42,7 @@
 }
 
 .news__item .news__meta-details {
+    width: 100%;
     position: absolute;
     top: 0;
     bottom: 0;
@@ -50,7 +52,6 @@
     background: rgba(0, 0, 0, 0.6);
     color: #fff;
     padding: 10px;
-    width: 100%;
     font-size: 0.9rem;
 }
 
@@ -125,6 +126,7 @@
 }
 
 @media (min-width: 640px) {
+
     .news__item {
         flex-direction: row;
         max-width: 700px;
@@ -166,4 +168,5 @@
     .news__item.alt .news__meta-details {
         padding-left: 25px;
     }
+
 }
