+++ news-JS/src/global.css
@@ -1,56 +1,66 @@
 @import url('https://fonts.googleapis.com/css2?family=Inter&family=Ubuntu:wght@700&display=swap');
+
 html,
 body {
+    height: 100%;
     color: #fff;
     background: #edf1ff;
     font-family: sans-serif;
-    height: 100%;
     margin: 0;
     padding: 0;
 }
+
 .wrapper {
     min-height: 100%;
     display: flex;
     flex-direction: column;
 }
+
 header {
     padding: 10px 30px;
 }
 
 header h1 {
+    color: #282d3c;
     font-size: 40px;
     font-family: 'Ubuntu', sans-serif;
-    color: #282d3c;
 }
+
 main {
     display: flex;
+    flex: 1 1 auto;
     flex-direction: column;
     justify-content: center;
     align-items: center;
-    flex: 1 1 auto;
 }
+
 footer {
     display: flex;
-    align-items: center;
     justify-content: space-between;
-    padding: 5px 5px 0 5px;
+    align-items: center;
     flex-wrap: wrap;
+    padding: 5px 5px 0 5px;
 }
+
 footer .copyright {
-    font-size: 20px;
     color: #333;
     text-align: center;
+    font-size: 20px;
     font-family: 'Inter', sans-serif;
 }
+
 footer .copyright a {
     color: #444;
 }
+
 footer .copyright a:hover {
     color: #555;
 }
+
 footer .copyright:before {
     content: '©';
 }
+
 footer .github-icon {
     width: 65px;
     height: 65px;
@@ -58,6 +68,7 @@ footer .github-icon {
     transition: 0.3s ease;
     margin-left: 10px;
 }
+
 footer .rs-icon {
     width: 180px;
     height: 55px;
@@ -65,7 +76,9 @@ footer .rs-icon {
     transform: scale(0.9);
     transition: 0.3s ease;
 }
+
 @media (max-width: 767px) {
+
     footer {
         flex-direction: column;
     }
@@ -74,7 +87,9 @@ footer .rs-icon {
         margin-left: 0px;
         margin-right: 0px;
     }
+
 }
+
 .rs-icon:hover,
 .github-icon:hover {
     transform: scale(1.1);
